requirements.txt
asgiref==3.3.1
astroid==2.4.2
beautifulsoup4==4.9.3
certifi==2020.12.5
chardet==4.0.0
colorama==0.4.4
coreapi==2.3.3
coreschema==0.0.4
dj-database-url==0.5.0
Django==3.1.4
django-audit-log==0.7.0
django-cors-headers==3.6.0
django-cors-middleware==1.5.0
django-htmlmin==0.11.0
djangorestframework==3.12.2
djangorestframework-jwt==1.11.0
html5lib==1.1
idna==2.10
isort==5.7.0
itypes==1.2.0
Jinja2==2.11.2
lazy-object-proxy==1.4.3
MarkupSafe==1.1.1
mccabe==0.6.1
Pillow==8.0.1
psycopg2==2.8.6
PyJWT==1.7.1
pylint==2.6.0
python-decouple==3.3
pytz==2020.5
requests==2.25.1
six==1.15.0
soupsieve==2.1
sqlparse==0.4.1
toml==0.10.2
uritemplate==3.0.1
urllib3==1.26.2
webencodings==0.5.1
whitenoise==5.2.0
wrapt==1.12.1
